backend/requirements-zero-cost.txt
# Optimized for Render Free Tier (512MB RAM)
# Excludes torch/transformers to use the built-in LocalEmbeddingModel fallback.

fastapi
uvicorn[standard]
python-multipart
python-dotenv
langchain
langchain-community
langchain-core
langchain-text-splitters
pypdf
pymupdf
reportlab
faiss-cpu
groq
sqlalchemy
psycopg2-binary
numpy
pandas
tqdm
pydantic
pydantic-settings
requests
